feat:add user management and profile page

+++ src/lib/format.ts
@@ -6,7 +6,7 @@ export function formatDate(
 
   try {
     return new Intl.DateTimeFormat('en-US', {
-      month: opts.month ?? 'long',
+      month: opts.month ?? 'short',
       day: opts.day ?? 'numeric',
       year: opts.year ?? 'numeric',
       ...opts
